+++ src/main/java/ba/rs/wabalabadubdub/model/Uloga.java
@@ -12,24 +12,24 @@ public class Uloga {
     @Id
     @GeneratedValue
     @Column(name="ID", unique = true)
-    private int ID;
+    private Long ID;
 
     @Column(name="naziv")
     private String naziv;
 
     public Uloga() {}
 
-    public Uloga(int id, String naziv) {
+    public Uloga(Long id, String naziv) {
         super();
         this.ID = id;
         this.naziv = naziv;
     }
 
-    public int getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
